Extract `testExpectTFatal()` to internal/plugintest package
